stage1/requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.15.0
accelerate>=0.25.0
jsonlines>=4.0.0
tqdm>=4.66.0
pyyaml>=6.0
numpy>=1.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Optional but recommended
flash-attn>=2.5.0
vllm>=0.2.6

# For data processing
requests>=2.31.0
beautifulsoup4>=4.12.0
